packages/gcp/pyproject.toml
[project]
name = "gcp-provider"
version = "0.1.0"
description = "GCP provider for Pragmatiks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pragmatiks-sdk",
    "google-cloud-secret-manager>=2.20.0",
]

authors = [
    { name = "Pragmatiks", email = "team@pragmatiks.io" },
]

[tool.uv.sources]
# During development, use the local SDK from the main monorepo
pragmatiks-sdk = { path = "../../../sdk", editable = true }

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.14.0",
]

[tool.pragma]
# Provider metadata - used by the platform for discovery and deployment
provider = "gcp"
package = "gcp_provider"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 120

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"
